### 2024-11-29 #676 @thomaszwagerman
`butterfly`: Verification of continually updating timeseries data where we expect new values, but want to ensure previous data remains unchanged.
Submitting Author Name: Thomas Zwagerman
Submitting Author Github Handle: @thomaszwagerman
Repository:  https://github.com/antarctica/butterfly
Version submitted: 1.0.0
Submission type: Standard

Version accepted: TBD

---

```
Package: butterfly
Title: Verification For Continually Updating Timeseries Data
Version: 1.0.0
    person("Thomas", "Zwagerman", , "thozwa@bas.ac.uk", role = c("aut", "cre"),
Description: Verification of continually updating timeseries data where we expect new values, but want to ensure previous data remains unchanged.
    cli,
    lifecycle,
    waldo
```

This package was written to handle the verification of continually updating time-series data, where we expect new values over time, but want to ensure previous data remains unchanged, and timesteps remain continuous. This package provides functionality that can be used as part of a data pipeline, to check and flag changes to previous data and prevent changes going unnoticed.

Researchers, data engineers, data stewards etc. We have implemented this package in an operational data pipeline, which extracts ERA5 data and performs some calculations to generate a new publishable dataset. [ERA5 can be subject to retrospective changes](https://cds.climate.copernicus.eu/datasets/reanalysis-era5-single-levels?tab=documentation), and so to prevent these changes going unnoticed and affecting our dataset, we use butterfly to verify our data has not unexpectedly changed compared to previously published versions. **New since pre-submission:** this package also contains functionality to check the continuity of timeseries data (due to instrument failure), and handle time series data with varying measurement frequencies.

(duplicated from pre-submission #665)

To my knowledge, there is no other package that handles the comparing of iterations of the same data, i.e. we want to verify there are no changes in previous data, but at the same time we do not want new data to throw an error and stop our pipeline.

[waldo](https://github.com/r-lib/waldo) - butterfly uses waldo::compare() in every function to provide a report on difference. There is therefore significant overlap, however butterfly builds on waldo by providing the functionality of comparing objects where we expect changes in some places (entirely new rows of data), but not in others (previously published data). butterfly also provides extra user feedback to provide clarity on what it is and isn’t comparing, due to the nature of comparing only “matched” rows.

[diffdf](https://github.com/gowerc/diffdf) - similar to waldo, but specifically for data frames, diffdf provides the ability to compare data frames directly. diffdf::diffdf() could have been used in our case, but I prefer waldo’s more explicit and clear user feedback. That said, there is significant overlap in functionality: butterfly::loupe() and diffdf::diffdf_has_issues() both provide a TRUE/FALSE difference check, while diffdf::diffdf_issue_rows() and butterfly::catch() both return the rows where changes have occurred. However, it lacks the flexibility of butterfly to compare object where we expect some changes, but not others.

[assertr](https://github.com/tonyfischetti/assertr) - assertr provides assertion functionality that can be used as part of a pipeline, and test assertions on a particular dataset, but it does not offer tools for comparison. I would use them hand-in-hand but butterfly has a comparison purpose.

[daquiri](https://github.com/ropensci/daiquiri/) - daquiri provides tools to check data quality and visually inspect timeseries data. It is a quality assurance package for timeseries, but does not have a comparison/verification purpose like butterfly.

N/A

#665

Since pre-submission, two new functions have been added to the package `timeline()` and `timeline_group()`. This has expanded the package to also deal with time series data from instruments/sensors, specifically those which are prone to error, or those which intentionally measure time stamped data at variable frequencies. Further feedback on controlling tolerance were also incorporated.

Another pre-submission suggestion, on checking relationships between variables was not implemented. I had a few passes at this, but I could not think of a generalised method of doing this effectively. I currently don't have a specific use case which suited, but I would be open to re-visiting and open to suggestions.

The only warning is "Function names are duplicated in other packages". Currently I've named functions in the way they made sense for this package, and following the "butterfly" theme. However, I recognise that in specific cases this might not be ideal. I would be very open to feedback/guidance from reviewers on best practices for naming these (with the personal preference of sticking with the "butterfly" theme), perhaps pre-pending functions as `bf_*()`? In the documentation I've tried to specify functions using butterfly::*().
