### 2024-12-04 #676 @emilyriederer
Hi @thomaszwagerman ! Thanks for your submission. We're excited to move forward with this review, and it looks like the package is in great shape. I will begin to look for a handling editor.
