### 2024-12-11 #676 @emilyriederer
---

Overall, the package looks in great shape for review. The level of documentation is especially impressive; I particularly like the graphics.

The repo also seems well organized, making the package easy to understand, install, and observe the development.

I marked the "Contributing information" box as complete since the package doesn't require any specific dev set up; however, you could consider adding explicit information on what types of contributions you welcome and how folks should engage (e.g. open an issue before beginning development? See example [here](https://github.com/ropensci/targets/blob/main/CONTRIBUTING.md) )

---
