### 2024-12-11 b030dfa
adding rOpenSci badge to README
