### 2024-12-12 b92b0a9
tailoring CONTRIBUTIONS.md to this package
