### 2024-12-12 2a0b3b8
make attribution text friendlier
