### 2024-12-12 8bbe65b
refer to ropensci guidance on news.md instead of tidyverse
