### 2024-12-12 a74a6ed
update NEWS.md with recent changes
