+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   articles/butterfly_in_pipeline: butterfly_in_pipeline.html
   articles/butterfly_paper: butterfly_paper.html
   butterfly: butterfly.html
-last_built: 2024-12-12T10:52Z
+last_built: 2024-12-12T10:54Z
 urls:
   reference: https://thomaszwagerman.github.io/butterfly/reference
   article: https://thomaszwagerman.github.io/butterfly/articles
