### 2024-12-12 #676 @thomaszwagerman
Thank you for your comments @emilyriederer - a very good point regarding being explicit about contributions. I have updated the CONTRIBUTING.md using the example you provided, and made it more bespoke to the package (antarctica/butterfly#30).

A small point of general clarification, before submitting for review I followed the [contribution guide](https://devguide.ropensci.org/maintenance_collaboration.html#contributing-guide), which states it is compulsory to have CONTRIBUTING.md under either `.github/` or `docs/`. Is this intended to discourage having the CONTRIBUTING.md at the top level? This is where I would probably expect it to be (I briefly forgot I already had one, under `./github/`), as it is in {targets}.
