### 2024-12-14 #676 @emilyriederer
@thomaszwagerman - let me introduce @qdread as our first reviewer! He's an expert reviewer with rOpenSci with lots of experience in longitudinal data.

@qdread - regarding the change in your email, please use the form above to update!
