### 2024-12-14 #676 @emilyriederer
@thomaszwagerman - great question and thanks for the catch! You're probably right its better in one of those locations to avoid confusion building the R package (otherwise we can tell the build to ignore it)
