### 2024-12-16 #676 @emilyriederer
@thomaszwagerman - I'm delighted to announce @TheAnalyticalEdge as our second review!

@TheAnalyticalEdge and @qdread - Please do not stress about the dates the bot "autoassigned" you. As we all discussed offline, the end of December is quite busy for all with the holidays. We can easily bump the date when the time comes.
