### 2024-12-29 #676 @qdread
My review is below. Happy New Year to all! And specifically for @emilyriederer Go Tar Heels!

- **Briefly describe any working relationship you have (had) with the package authors.** *None to declare.*

- [x] **Performance:** Any performance claims of the software have been confirmed. *Not applicable.*

---

I enjoyed reviewing this package. Thanks for giving me the opportunity to take a look at it! Below I have written some elaborations on the checklist items that I did not mark complete, and point-by-point comments on different things I noticed while looking over the package. I would say that most of these comments are more like suggestions for possible improvements, than me noticing fatal flaws that must be fixed or else. In general, the package does a well-defined job, does it correctly, and documents itself well.

One general comment I have is that the code execution seemed fairly slow for the examples that only use input datasets with a few dozen rows. I would be concerned that it would scale poorly to large datasets. It might be useful to test this with (realistically) large input datasets to see if the performance is still acceptable.

#### Explanation of checklist items not marked complete

- **Community guidelines**: There is no explicit mention of how to contribute in README and there is no CONTRIBUTING. This is easy to fix. Just add a few sentences to the readme saying that would-be contributors should submit a well-documented pull request, or something along those lines.
- **Automated tests**: The tests that are currently in the test suite are all great. They all pass when run on my local machine and have 100% code coverage (I verified this by running `covr::report()` locally). That is excellent. What I think is potentially missing from the test suite are some tests for what the functions return if passed invalid or incorrectly formatted input. All the tests currently use `butterflycount` which is good to demonstrate that expected output is returned if passing clean well-formatted input. It would be good to make sure that appropriate errors, warnings, etc., are returned if the input isn't so clean.

#### Specific comments: suggested improvements to the code

- It is good that verbosity of the output can be set globally with `options()` but it might also be helpful to include an argument to each individual function allowing `verbose = FALSE` to be set on an individual basis.
- I don't quite understand the purpose of `loupe()` when `catch()` exists. When I run `loupe()` I see which rows, if any, have changed, but they just appear printed out and can't be assigned to a data frame. What is the usefulness of `loupe()` if `catch()` gives me the same side effect and also returns something useful? It would almost make more sense to combine those two functions into one, that could both display the changed rows or return them. A nifty way to do this might be to make `catch()` return an object with a special class that inherits from `data.frame`, that has its own special print method that would show the nice informative messages about changed rows, but could still be used as a data frame if needed.
- One other minor suggestion is that in `loupe()`, the messages are slightly confusing when no rows are new. For changed rows, it makes sense because if rows have changed, a message that rows have changed is followed by a list of changed rows. If no rows have changed, we see `And there are no differences with previous data.` But regardless of whether there are new rows or not, the message is printed `The following rows are new in 'df_current': ` . I would recommend replacing this with `There are no new rows in 'df_current'` if no rows are new.

#### Specific comments: vignettes

- The main vignette doesn't really start with an intro that clearly explains what the package is doing. The brief intro sentences may be slightly confusing. I would recommend adding another sentence or two explaining in big-picture terms what are the inputs you would need, and what outputs/results you should expect to get.
- The graphics on the main vignette showing what the package does are very nice!
- The operational data pipeline vignette was a nice read. It really helped illustrate the utility of the package. Again, I liked the flowchart graphics a lot. The only minor comment I have is that I would note somewhere that the scripts are run in a Bash shell. It says this on the header of one of the scripts but it may not be obvious to people who are not familiar with shell scripting.
- The title of the third vignette is very long, so it doesn't display well in the "Articles" dropdown menu on the pkgdown site. I noticed this is an unfinished article so I wasn't sure why it is already published on the public-facing site to begin with. I did not review it very closely because it seemed unfinished.
- I noticed a few typo errors in the vignettes (grammar, punctuation). They might need to be checked over.

#### Specific comments: function documentation

- The Description field for `loupe()` explains the function's name but it does not say what the function does. Because the first sentence after the title becomes the Description field in the help documentation, why not switch sentence 1 and 2?
- Either documentation or vignette should say somewhere that `waldo::compare()` is an expanded version of `all.equal()`. More people are familiar with that base R function. At least for me, I did not know about `compare()` but I understood better what it was doing when I realized it is sort of a fancier version of `all.equal()`.
- The documentation of `catch()`, line 52, says inner join, but I think it should say anti join.
- The naming of `timeline()` and `timeline_group()` is not really consistent with the other functions. The names aren't that descriptive of what they do. (And unfortunately they do not have a cool lepidopteran-themed name!) I might recommend changing those names, of course keeping the old ones as deprecated for backward compatibility.
