### 2025-01-08 #676 @thomaszwagerman
Happy new year all!

@qdread I am very grateful for your very thorough and thoughtful review. All excellent points raised - I will look at these and address them in detail in the coming weeks. In the meantime I just wanted to say thank you!
