### 2025-01-08 antarctica/butterfly#31 @thomaszwagerman
Test performance with larger datasets
See ropensci/software-review#676

Reviewer reported slow performance with standard dataset, test if this could become problematic with larger data
