### 2025-01-08 antarctica/butterfly#32 @thomaszwagerman
Add contributing sections to README
See ropensci/software-review#676
