### 2025-01-08 antarctica/butterfly#33 @thomaszwagerman
Tests for invalid/incorrectly formatted input
Improve testing suite to include tests for incorrectly formatted datasets as input, what should each function return?

See ropensci/software-review#676
