### 2025-01-08 antarctica/butterfly#34 @thomaszwagerman
Improve `loupe()` feedback when no new rows
From the reviewer:

> For changed rows, it makes sense because if rows have changed, a message that rows have changed is followed by a list of changed rows. If no rows have changed, we see And there are no differences with previous data. But regardless of whether there are new rows or not, the message is printed The following rows are new in 'df_current':  . I would recommend replacing this with There are no new rows in 'df_current' if no rows are new.

See ropensci/software-review#676
