### 2025-01-08 antarctica/butterfly#35 @thomaszwagerman
Vignette improvements
Reviewer comments on vignettes:

- [x] Impove main vignette introduction. From the reviewer: "The main vignette doesn't really start with an intro that clearly explains what the package is doing. The brief intro sentences may be slightly confusing. I would recommend adding another sentence or two explaining in big-picture terms what are the inputs you would need, and what outputs/results you should expect to get."
- [x] Mention scripts are run in Bash shell.
- [x] Hide JOSS vignette until it is ready.
- [x] Grammar and punctuation check.

See ropensci/software-review#676
