### 2025-01-08 antarctica/butterfly#36 @thomaszwagerman
Documentation improvements
From reviewer comments:

- [x] Specify what loupe does in the description.
- [x] Elaborate on `all.equal()`, in addition to `waldo::compare()`
- [x] `catch()` line 52, replace `inner_join()` with `anti_join()`

Other comments included `timeline()` and `timeline_group()` renaming, I will cover these separately - pending possible further function naming comments from the other reviewer.

See ropensci/software-review#676
