### 2025-01-10 b7e7971
add contributing sections to readme
