+++ README.md
@@ -4,6 +4,7 @@
 # butterfly <a href="https://thomaszwagerman.github.io/butterfly/"><img src="man/figures/logo.png" align="right" height="139" alt="butterfly website" /></a>
 
 <!-- badges: start -->
+
 [![R-CMD-check](https://github.com/thomaszwagerman/butterfly/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/thomaszwagerman/butterfly/actions/workflows/R-CMD-check.yaml)
 [![Codecov test
 coverage](https://codecov.io/gh/thomaszwagerman/butterfly/branch/main/graph/badge.svg)](https://app.codecov.io/gh/thomaszwagerman/butterfly?branch=main)
@@ -13,7 +14,6 @@ stable](https://img.shields.io/badge/lifecycle-stable-brightgreen.svg)](https://
 state and is being actively
 developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 [![pkgcheck](https://github.com/thomaszwagerman/butterfly/workflows/pkgcheck/badge.svg)](https://github.com/thomaszwagerman/butterfly/actions?query=workflow%3Apkgcheck)
-[![Status at rOpenSci Software Peer Review](https://badges.ropensci.org/676_status.svg)](https://github.com/ropensci/software-review/issues/676)
 <!-- badges: end -->
 
 The goal of butterfly is to aid in the verification of continually
@@ -375,3 +375,17 @@ Other functions include `all.equal()` (base R) or
 Read more about how `butterfly` is [used in an operational data
 pipeline](https://thomaszwagerman.github.io/butterfly/articles/butterfly_in_pipeline.html)
 to verify a continually updated **and** published dataset.
+
+## Contributing
+
+### Without write access
+
+Corrections, suggestions and general improvements are welcome as issues.
+
+You can also suggest changes by forking this repository, and opening a
+pull request. Please target your pull requests to the main branch.
+
+### With write access
+
+You can push directly to main for small fixes. Please use PRs to main
+for discussing larger updates.
