### 2025-01-10 af3b33d
clarify the purpose of loupe, highlight the difference with catch
