### 2025-01-10 350da55
expand on waldo compare, mention base all.equal
