+++ R/loupe.R
@@ -11,7 +11,9 @@
 #' (usually "time" or "datetime in a timeseries).
 #'
 #' It informs the user of new (unmatched) rows which have appeared, and then
-#' returns a `waldo::compare()` call to give a detailed breakdown of changes.
+#' returns a `waldo::compare()` call to give a detailed breakdown of changes. If
+#' you are not familiar with `waldo::compare()`, this is an expanded and more
+#' verbose function similar to base R's `all.equal()`.
 #' 
 #' `loupe()` will then return TRUE if there are not changes to previous data,
 #' or FALSE if there are unexpected changes. If you want to extract changes as
