### 2025-01-10 9e88ac7
mention that shell scripts are ran in the bash shell
