### 2025-01-10 thomaszwagerman/butterfly#35 @thomaszwagerman
Hiding vignettes is not really possible, like it is with draft quarto pages
