### 2025-01-14 78260fe
more detailed explanation of what butterfly is for, and what it expects
