### 2025-01-14 08b7556
style and grammatical updates
