### 2025-01-14 thomaszwagerman/butterfly#34 @thomaszwagerman
When there are no new rows, it returns "There are no new rows. Check 'df_current' is your most recent data, and 'df_previous' is your previous data. If comparing directly, try waldo::compare()".

I think this already satisfies the requirements.
