+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   articles/butterfly_in_pipeline: butterfly_in_pipeline.html
   articles/butterfly_paper: butterfly_paper.html
   butterfly: butterfly.html
-last_built: 2024-12-12T10:54Z
+last_built: 2025-01-14T11:33Z
 urls:
   reference: https://thomaszwagerman.github.io/butterfly/reference
   article: https://thomaszwagerman.github.io/butterfly/articles
