### 2025-01-14 thomaszwagerman/butterfly#31 @thomaszwagerman
![Screenshot from 2025-01-14 12-02-03](https://github.com/user-attachments/assets/f541a08b-8325-455f-852a-a0040e19e960)

The above is `loupe()` performance on datasets with 1638100 and 1638000 rows. Performance is not great, but acceptable for an action that is likely to be ran as a one-off.

The main culprit is the underlying use of `waldo::compare()`, rather than anything introduced by butterfly itself. I think therefore, no further action required as part of **this** package.
