### 2025-01-19 #676 @emilyriederer
Bumping for @TheAnalyticalEdge . I know you noted that you wouldn't have time to review until the new year, but do you have a sense when you may be ready?
