### 2025-02-02 #676 @maurolepore
Dear @thomaszwagerman and @emilyriederer this is just to mark the start of my EiC rotation and record what I see. I see one review is delayed but not by much and the editor is following up. I'll step back and let things flow. Thanks!
