### 2025-02-24 #676 @TheAnalyticalEdge
## Package Review

-   **Briefly describe any working relationship you have (had) with the package authors.** None to declare.
-   ☒ As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

#### Documentation

-   ☒ **A statement of need:** clearly stating problems the software is designed to solve and its target audience in README
-   ☒ **Installation instructions:** for the development version of package and any non-standard dependencies in README
-   ☒ **Vignette(s):** demonstrating major functionality that runs successfully locally
-   ☒ **Function Documentation:** for all exported functions
-   ☒ **Examples:** (that run successfully locally) for all exported functions
-   ☒ **Community guidelines:** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).

#### Functionality

-   ☒ **Installation:** Installation succeeds as documented.
-   ☒ **Functionality:** Any functional claims of the software have been confirmed.
-   ☒ **Performance:** Any performance claims of the software have been confirmed.
-   ☒ **Automated tests:** Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.
-   ☒ **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines.

Estimated hours spent reviewing: 2-3 hours

-   ☒ Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer (“rev” role) in the package DESCRIPTION file.

### Review Comments

Hello. Sincere apologies this took me far long than I had hoped to complete. Thank you for the opportunity to look at this package though, I do a bit with time series data and can see how this package is niche but very useful given differences from other packages available.

Minor comments(and some statements)

-   nice use of the word ‘loupe’ to be a function name

-   the loupe function is good. The example highlights a single change, but when I manipulated some data, it picks up and clearly highlights all changes and I thought the layout was good.

-   however, with loupe, the ‘yellow’ colour is used for text ‘The following values have changes from the previous data’ and then in the output `old$count` and `new$count` is green? I thought should remain yellow, because I’d imagine most people associate green with good? and would be consistent between text and output.

-   with ?timeline, please clarify in the Details section of the help file, if the expected lag is based on the smallest units used, one can specify a weekly lag by having YYY-MM-DD data, but then specifying expected\_lag of 7. (That is what I believe is going on, but the example for monthly data is really clear, and wanted to double check here)

-   ?timeline\_group is also a nice function. Well done.
