### 2025-02-25 antarctica/butterfly#38 @thomaszwagerman
Check loupe feedback colour for consistency.
From ropensci/software-review#676:

> with loupe, the ‘yellow’ colour is used for text ‘The following values have changes from the previous data’ and then in the output old$count and new$count is green? I thought should remain yellow, because I’d imagine most people associate green with good? and would be consistent between text and output.

Check colours for consistency and update where required. This may be part of `waldo`, however.
