### 2025-02-25 antarctica/butterfly#39 @thomaszwagerman
Clarify timeline details for expected lag units
From ropensci/software-review#676:

> ?timeline, please clarify in the Details section of the help file, if the expected lag is based on the smallest units used, one can specify a weekly lag by having YYY-MM-DD data, but then specifying expected_lag of 7.

Check the expected lag units descriptions and update for clarity on the lag units.
