### 2025-02-27 thomaszwagerman/butterfly#38 @thomaszwagerman
I agree with this comment. Unfortunately the colour of the output (not the message) is controlled by `waldo::compare()`, and therefore I cannot change this.
