+++ README.md
@@ -14,7 +14,6 @@ stable](https://img.shields.io/badge/lifecycle-stable-brightgreen.svg)](https://
 state and is being actively
 developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 [![pkgcheck](https://github.com/thomaszwagerman/butterfly/workflows/pkgcheck/badge.svg)](https://github.com/thomaszwagerman/butterfly/actions?query=workflow%3Apkgcheck)
-[![Status at rOpenSci Software Peer Review](https://badges.ropensci.org/676_status.svg)](https://github.com/ropensci/software-review/issues/676)
 <!-- badges: end -->
 
 The goal of butterfly is to aid in the verification of continually
@@ -94,6 +93,8 @@ demonstrate butterfly functionality:
   fictional butterfly counts for a given date.
 - `forestprecipitation` - a list of monthly dataframes, which contain
   fictional daily precipitation measurements for a given date.
+- `butterflymess` - a messy version of `butterflycount`, provided for
+  testing purposes.
 
 ## Examples
 
