### 2025-03-04 ab8899c
update NEWS.md
