### 2025-03-04 170d5db
update description and metadata for new release
