+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: butterfly
 Title: Verification For Continually Updating Timeseries Data
-Version: 1.0.0
+Version: 1.1.0
 Authors@R: 
     person("Thomas", "Zwagerman", , "thozwa@bas.ac.uk", role = c("aut", "cre"),
            comment = c(ORCID = "0009-0003-3742-3234"))
