### 2025-03-04 50d8a7d
restructuring of paper, incorporate pipeline article where relevant
