+++ vignettes/articles/butterfly_paper.Rmd
@@ -29,55 +29,86 @@ knitr::opts_chunk$set(
 
 # Summary
 
-Previously recorded data could be revised after initial publication number of reasons, such as discovery of an inconsistency or error, a change in methodology or instrument re-calibration. When using other data to generate your own, it is crucial to reference the exact version of the data used, in order to maintain data provenance. Unnoticed changes in previous data could have unintended consequences, such as invalidating a published dataset’s Digital Object Identfier (DOI), or altering future predictions if used as input in forecasting models.
+Unnoticed changes in data streams or published datasets can have unintended consequences.
 
-But what if you are not aware of upstream changes to your input data? Monitoring data sources for these changes is not always possible. Here we present butterfly, an R package for the verification of continually updating timeseries data where we expect new values, but want to ensure previous data remains unchanged.
+Environmental sensor behaviour during failure, calibration, or software updates is not consistent across sensors. This can affect measurement frequency or jumps in time. At the same time, some sensors are intentionally designed to have variable measurement frequency.
+
+Other unnoticed changed in data can occur in previously published data. These can be revised after initial publication for a number of reasons, such as discovery of an inconsistency or error, a change in methodology or instrument re-calibration. Unintended consequences include Digital Object Identifier (DOI) invalidation, or altering future predictions if used as input in forecasting models. This is especially relevant when using other data to generate your own. To maintain data provenance, it is crucial to reference the exact version of the data used.
+
+But what if you are not aware of upstream changes to your input data? Monitoring data sources for these changes is not always possible. Here we present butterfly, an R package for the verification of continually updating timeseries data where we expect new values, but want to ensure previous data remains unchanged, and measurement frequency is maintained as expected.
 
 The intention of butterfly is to check for changes in previously published data, and warn the user with a report that contains as much details as possible. This will allow them to stop unintended data transfer, revise their published data, release a new version and communicate the significance of the change to their users.
 
 # Statement of Need
 
-Semantic versioning is widely adopted in research software [(Preston-Werner 2013)](https://semver.org/spec/v2.0.0.html), but as outlined above datasets may also change for any number of reasons. It is therefore important to cite the exact extract of data you are using in your research to maintain reproducibility [(Klump et al. 2021)](https://datascience.codata.org/articles/10.5334/dsj-2021-012). It is not only important to indicate to users that there has been a change, but also **what** that change is. 
+Semantic versioning is widely adopted in research software [(Preston-Werner 2013)](https://semver.org/spec/v2.0.0.html), but as outlined above a published dataset can also change for a number of reasons. It is therefore important to cite the exact version of data you are using in your research to maintain reproducibility [(Klump et al. 2021)](https://datascience.codata.org/articles/10.5334/dsj-2021-012). If there are any changes this should be indicated to users, with an explanation of what that change is, and why it is significant. 
 
-This may be especially relevant for Information Management Frameworks for Digital Twins (Siddorn et al. 2022). A digital twin might rely on any number of source data, whether live sensor streams or environmental forecasting models. To achieve a FAIR implementation (Wilkinson et al. 2016) of a Digital Twin, data provenance must be maintained, clearly documented for users and available in machine-readable format.
[…]
