### 2025-03-04 293890a
formatted references in paper.bib
