+++ vignettes/articles/paper.bib
@@ -1,0 +1,130 @@
+@article{klump_2021,
+    url = {https://datascience.codata.org/articles/1246/files/submission/proof/1246-1-9224-1-10-20210323.pdf},
+    year = {2021},
+    title = {Versioning Data Is About More than Revisions: A Conceptual Framework and
+    Proposed Principles},
+    author = {Klump, J., Wyborn, L., Wu, M., Martin, J., Downs, R.R., Asmi, A.},
+    number = {1},
+    volume = {20},
+    journal = {Data Science Journal}}
+
+@misc{preston-werner_2024,
+    note = {Accessed: 2024-10-28},
+    year = {2024},
+    title = {Semantic Versioning 2.0.0 | Semantic Versioning},
+    author = {Preston-Werner, T.},
+    howpublished = {\url{https://semver.org/spec/v2.0.0.html}}}
+
+@article{blair_2023,
+    doi = {10.1002/env.2789},
+    url = {http://dx.doi.org/10.1002/env.2789},
+    issn = {1099-095X},
+    year = {2023},
+    month = {jan},
+    title = {The role of data science in environmental digital twins: In praise of the
+    arrows},
+    author = {Blair, G.S. and Henrys, P.A.},
+    number = {2},
+    volume = {34},
+    journal = {Environmetrics},
+    publisher = {Wiley}}
+
+@article{afanasyev_2015,
+    doi = {10.1371/journal.pone.0126292},
+    url = {http://dx.doi.org/10.1371/journal.pone.0126292},
+    issn = {1932-6203},
+    year = {2015},
+    month = {apr},
+    pages = {e0126292},
+    title = {Increasing Accuracy: A New Design and Algorithm for Automatically Measuring
+    Weights, Travel Direction and Radio Frequency Identification (RFID) of
+    Penguins},
+    author = {Afanasyev, V., Buldyrev, S.V., Dunn, M.J., Robst, J., Preston, M., Bremner,
+    S.F., Briggs, D.R., Brown, R., Adlard, S., Peat, H.J.},
+    editor = {Humphries, Stuart},
+    number = {4},
+    volume = {10},
+    journal = {PLOS ONE},
+    publisher = {Public Library of Science (PLoS)}}
+
+@article{hersbach_2020,
+    doi = {10.1002/qj.3803},
+    url = {http://dx.doi.org/10.1002/qj.3803},
+    issn = {1477-870X},
+    year = {2020},
+    month = {jun},
+    pages = {1999â2049},
+    title = {The ERA5 global reanalysis},
+    author = {Hersbach, H., Bell, B., Berrisford, P., Hirahara, S., HorÃ¡nyi, A., and
+    others},
+    number = {730},
+    volume = {146},
+    journal = {Quarterly Journal of the Royal Meteorological Society},
+    publisher = {Wiley}}
+
+@article{hosking_2016,
+    doi = {10.1002/2015gl067143},
+    url = {https://doi.org/10.1002/2015gl067143},
+    issn = {0094-8276},
+    year = {2016},
+    month = {Jan},
+    pages = {367--376},
+    title = {Future circulation changes off West Antarctica: Sensitivity of the Amundsen
+    Sea Low to projected anthropogenic forcing},
+    author = {Hosking, J.S., Orr, A., Bracegirdle, T.J., Turner, J.},
+    number = {1},
+    volume = {43},
+    journal = {Geophysical Research Letters},
+    publisher = {American Geophysical Union (AGU)}}
+
+@misc{siddorn_2022,
+    doi = {10.5281/ZENODO.7004351},
+    url = {https://zenodo.org/record/7004351},
+    year = {2022},
[…]
