+++ vignettes/articles/butterfly_paper.Rmd
@@ -5,7 +5,7 @@ tags:
 - quality assurance
 - timeseries
 - ERA5
-date: "23 October 2024"
+date: "04 March 2025"
 affiliations:
 - name: British Antarctic Survey, UK
   index: 1
@@ -14,6 +14,7 @@ authors:
   orcid: "0009-0003-3742-3234"
   equal-contrib: true
   affiliation: 1
+bibliography: paper.bib
 ---
 
 ```{r, include = FALSE}
@@ -23,8 +24,6 @@ knitr::opts_chunk$set(
 )
 ```
 
-#\> left out bibliography: paper.bib from yaml
-
 # **Active work in progress!**
 
 # Summary
@@ -33,7 +32,7 @@ Unnoticed changes in data streams or published datasets can have unintended cons
 
 Environmental sensor behaviour during failure, calibration, or software updates is not consistent across sensors. This can affect measurement frequency or jumps in time. At the same time, some sensors are intentionally designed to have variable measurement frequency.
 
-Other unnoticed changed in data can occur in previously published data. These can be revised after initial publication for a number of reasons, such as discovery of an inconsistency or error, a change in methodology or instrument re-calibration. Unintended consequences include Digital Object Identifier (DOI) invalidation, or altering future predictions if used as input in forecasting models. This is especially relevant when using other data to generate your own. To maintain data provenance, it is crucial to reference the exact version of the data used.
+Other unnoticed changed in data can occur in previously published data. These can be revised after initial publication, due to discovery of an inconsistency or error, a change in methodology or instrument re-calibration. Unintended consequences include Digital Object Identifier (DOI) invalidation, or altering future predictions if used as input in forecasting models. This is especially relevant when using other data to generate your own. To maintain data provenance, it is crucial to reference the exact version of the data used.
 
 But what if you are not aware of upstream changes to your input data? Monitoring data sources for these changes is not always possible. Here we present butterfly, an R package for the verification of continually updating timeseries data where we expect new values, but want to ensure previous data remains unchanged, and measurement frequency is maintained as expected.
 
@@ -41,11 +40,11 @@ The intention of butterfly is to check for changes in previously published data,
 
 # Statement of Need
 
-Semantic versioning is widely adopted in research software [(Preston-Werner 2013)](https://semver.org/spec/v2.0.0.html), but as outlined above a published dataset can also change for a number of reasons. It is therefore important to cite the exact version of data you are using in your research to maintain reproducibility [(Klump et al. 2021)](https://datascience.codata.org/articles/10.5334/dsj-2021-012). If there are any changes this should be indicated to users, with an explanation of what that change is, and why it is significant. 
+Semantic versioning is widely adopted in research software [@preston-werner_2024], but as outlined above a published dataset can also change for a number of reasons. It is therefore important to cite the exact version of data you are using in your research to maintain reproducibility [@klump_2021]. If there are any changes this should be indicated to users, with an explanation of what that change is, and why it is significant. 
 
[…]
