### 2025-03-05 291214e
adding plots to illustrate sensor story, these are placeholders
