+++ vignettes/articles/paper.bib
@@ -3,7 +3,7 @@
     year = {2021},
     title = {Versioning Data Is About More than Revisions: A Conceptual Framework and
     Proposed Principles},
-    author = {Klump, J., Wyborn, L., Wu, M., Martin, J., Downs, R.R., Asmi, A.},
+    author = {{Klump}, J. and {Wyborn}, L. and {Wu}, M. and {Martin}, J. and {Downs}, R.R. and {Asmi}, A.},
     number = {1},
     volume = {20},
     journal = {Data Science Journal}}
@@ -23,7 +23,7 @@
     month = {jan},
     title = {The role of data science in environmental digital twins: In praise of the
     arrows},
-    author = {Blair, G.S. and Henrys, P.A.},
+    author = {Blair, G.S., Henrys, P.A.},
     number = {2},
     volume = {34},
     journal = {Environmetrics},
