### 2025-03-10 #676 @emilyriederer
Thanks, @thomaszwagerman for the meticulous reply!

@qdread @TheAnalyticalEdge - could you please review Thomas' comments and see what you think? When you're ready, you may use the [approval template](https://devguide.ropensci.org/approval2template.html).
