### 2025-03-22 #676 @ropensci-review-bot
Approved! Thanks @thomaszwagerman for submitting and @qdread, @TheAnalyticalEdge for your reviews! :grin:
