### 2025-03-24 #676 @thomaszwagerman
Thank you @emilyriederer @qdread @TheAnalyticalEdge for the great comments and reviews, this was a very pleasant process!

Please let me know if you consent to being added as "rev"-type contributors in the DESCRIPTION, and I would be more than happy to add you. Many thanks!
